Move yeelight services to component domain
